Refactor: 변수명, 컴포넌트명 알기쉽게 변경
