+++ src/components/Footer/footerListData.js
@@ -1,5 +1,6 @@
 const footerListData = [
   {
+    id: 'footerNum1',
     title: 'COMPANY',
     list: [
       '한눈에 보기',
@@ -9,25 +10,24 @@ const footerListData = [
       '세계의 스타벅스',
       '글로벌 뉴스룸',
     ],
-    id: 'footerNum1',
   },
   {
+    id: 'footerNum2',
     title: 'CORPORATE SALES',
     list: ['단체 및 기업 구매 안내'],
-    id: 'footerNum2',
   },
   {
+    id: 'footerNum3',
     title: 'PARTNERSHIP',
     list: ['신규 입점 제의', '협력 고객사 등록 신청'],
-    id: 'footerNum3',
   },
   {
+    id: 'footerNum4',
     title: 'ONLINE COMMUNITY',
     list: ['페이스북', '트위트', '유튜브', '블로그', '인스타그램'],
-    id: 'footerNum4',
   },
-  { title: 'RECRUIT', list: ['채용 소개', '채용 지원하기'], id: 'footerNum5' },
-  { title: 'WEBUKS', list: [''], id: 'footerNum6' },
+  { id: 'footerNum5', title: 'RECRUIT', list: ['채용 소개', '채용 지원하기'] },
+  { id: 'footerNum6', title: 'WEBUKS', list: [''] },
 ];
 
 export default footerListData;
