--- src/pages/wookchanglee/Detail/EachReview/EachReview.scss
@@ -1,36 +1,0 @@
-.EachReview {
-  display: flex;
-  justify-content: space-between;
-  margin: 6px 0;
-
-  .content {
-    display: flex;
-
-    .username {
-      margin-right: 10px;
-      font-weight: 700;
-    }
-  }
-
-  .option {
-    display: flex;
-
-    .like {
-      position: relative;
-      top: -4px;
-      margin: 0 0 10px 10px;
-      color: #ccc;
-      font-size: 18px;
-      cursor: pointer;
-
-      .fas {
-        color: #f5425d;
-      }
-    }
-
-    .delete {
-      margin-left: 15px;
-      cursor: pointer;
-    }
-  }
-}
