+++ src/pages/wookchanglee/Detail/Reviews/Reviews.scss
@@ -1,0 +1,32 @@
+.Reviews {
+  h2 {
+    margin: 40px 0 10px 0;
+    padding-bottom: 10px;
+    border-bottom: 1px solid #ccc;
+    font-size: 20px;
+    font-weight: 600;
+  }
+
+  .reviewList {
+    margin: 20px 0;
+  }
+
+  .inputBox {
+    display: flex;
+
+    input {
+      height: 30px;
+      padding-left: 10px;
+      border: 0;
+      background-color: #efefef;
+    }
+
+    .id {
+      flex: 1;
+    }
+
+    .content {
+      flex: 5;
+    }
+  }
+}
