--- src/styles/_common.scss
@@ -1,6 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-a {
-  text-decoration: none;
-}
